Refactored

+++ src/schemas/movies.py
@@ -15,8 +15,8 @@ class MovieDetailResponseSchema(BaseModel):
     orig_title: str
     status: str
     orig_lang: str
-    budget: int
-    revenue: int
+    budget: float
+    revenue: float
     country: str
 
     model_config = ConfigDict(from_attributes=True)
